djangoProject/requirements.txt
django
paramiko
djangorestframework
django-cors-headers